Reutilizing this sidescroller engine for overworld RPG
This is a pre-emptive commit beforeI start adding in animations. You'll see a few tools already in the project to make this happen. Map is made with Tiled and parsed with STI. Camera via Gamera, player movement and collisions are down (at least with platforms) and that's about it (there's probably more I'm missing).